Fix segfault when trying to plot.
Needed to update function calls from show() to draw() with the change to
Python 3.x.  Also the NavigationToolbar2TkAgg() method has been renamed
to NavigationToolbar2Tk().

The plots are ugly (at least on a Mac) with the axes text being to large
and overrunning, and the plot is not fully displayed, but the program
no longer segfaults.